requirements.txt
mistralai==2.0.0
python-dotenv==1.2.2
python-telegram-bot==22.0
httpx>=0.28.1
requests==2.33.1